requirements.txt
psycopg2-binary>=2.7.4
boto3>=1.6.22
requests>=2.18.4